main: return Status instead of raw isize

+++ src/uefi_entry/mod.rs
@@ -7,5 +7,5 @@ pub extern "win64" fn efi_entry(image_handle: uefi::Handle,
                                 system_table: *const uefi::SystemTable)
                                 -> isize {
     uefi::set_system_table(system_table).console().reset();
-    ::efi_main()
+    ::efi_main() as isize
 }
